requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
banal==1.0.6
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataset==1.6.0
datasets==2.11.0
dbus-python==1.2.18
defer==1.0.6
dill==0.3.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.2
greenlet==2.0.2
gyp==0.1
httplib2==0.20.2
huggingface-hub==0.14.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
multiprocess==0.70.14
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
packaging==23.1
pandas==2.0.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.2.0
protobuf==3.12.4
ptyprocess==0.7.0
pyarrow==11.0.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.8
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
responses==0.18.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.47
sqlparse==0.4.4
stripe==5.4.0
systemd-python==234
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.22.0
wadllib==1.3.6
xdg==5
xkit==0.0.0
xxhash==3.2.0
yarl==1.9.2
zipp==1.0.0
